+++ CHANGELOG.md
@@ -6,11 +6,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [4.5.1] - 2024-02-03
 
 ### Fixed
 
 - LSP: Notify if an LSP request was made but no rust-analyzer client is attached.
+- Neotest: Only the current buffer was queried for test positions.
 
 ## [4.5.0] - 2024-02-02